+++ src/main/java/org/example/competition22/logic/WallsChecker.java
@@ -7,12 +7,12 @@ import org.example.competition22.data.MoveRequest;
 import java.util.Map;
 
 public class WallsChecker {
-    public static void check(MoveRequest request, Map<Direction, Integer> directions) {
+    public static void check(MoveRequest request, Map<Direction, Double> directions) {
         for (Direction direction : Direction.values()) {
             var possibleMove = Coordinate.getNextCoordinate(request.you.head, direction);
             if (possibleMove.x < 0 || possibleMove.x == request.board.width
                     || possibleMove.y < 0 || possibleMove.y == request.board.height) {
-                directions.put(direction, -1);
+                directions.put(direction, -1.0);
             }
         }
     }
